requirements.txt
tensorflow
pandas
dvc
notebook
numpy
matplotlib
seaborn
pyYAML
tqdm
ensure
python-box
joblib
types-PyYAML
scipy
Flask
Flask-Cors
-e .